Former-commit-id: a47a232b93b04244f02c604f6670d1bda48fecda

+++ hq-server/src/main/java/org/hyperic/hq/inventory/domain/Config.java
@@ -141,7 +141,7 @@ public class Config {
         return entityManager().createQuery("select o from Config o", Config.class).getResultList();
     }
 
-    public static Config findById(Long id) {
+    public static Config findById(Integer id) {
         if (id == null)
             return null;
         return entityManager().find(Config.class, id);
